2003-12-14 Anton Blanchard <anton@samba.org>
	* syscall.c (get_error) [POWERPC]: Cast result to unsigned long for
	ppc64, we were misreporting syscalls that really succeeded as having
	failed.

+++ syscall.c
@@ -1414,7 +1414,7 @@ struct tcb *tcp;
 		}
 #else
 #ifdef POWERPC
-		if (result && (unsigned) -result < nerrnos) {
+		if (result && (unsigned long) -result < nerrnos) {
 			tcp->u_rval = -1;
 			u_error = -result;
 		}
